GETメソッドのディレクトリを変更して確認画面で正しく目標を表示できるようにする
