+++ frontend/app/api/student/goals/route.ts
@@ -24,26 +24,3 @@ export async function POST(req: Request) {
     return NextResponse.json({ message: "INTERNAL_SERVER_ERROR"}, { status: 500})
   }
 }
-
-export async function GET (req: Request, { params }: { params: { goalId: string } }) {
-  try {
-    const goalId = Number(params.goalId)
-
-    const { status, data, setCookie } = await railsFetch(`/api/v1/student/goals/${goalId}`, {
-      method: "GET"
-    })
-
-    const nextResponse = NextResponse.json(data, { status })
-
-    if (setCookie) nextResponse.headers.set('set-cookie', setCookie)
-
-    return nextResponse
-
-  } catch (error) {
-    if (error instanceof RailsUnauthorizedError) {
-        return NextResponse.json({ message: "UNAUTHORIZED" }, { status: 401 })
-    }
-
-    return NextResponse.json({ message: "INTERNAL_SERVER_ERROR" }, { status: 500 })
-  }
-}
